backend/requirements.txt
# Core API framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Settings & validation
pydantic==2.7.1
pydantic-settings==2.2.1

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Auth
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
boto3==1.34.101

# Image processing
Pillow==10.3.0
numpy==1.26.4

# TFLite inference. On Windows we use full TensorFlow because tflite-runtime
# doesn't ship Windows wheels. In production (Linux), we'll switch back.
tensorflow==2.19.0

# Gemini API for treatment advice generation
google-generativeai==0.7.2

# HTTP client + Redis
httpx==0.27.0
redis==5.0.4